testshark/requirements.txt
requests
pygments
django-allauth
javaparser  # For Java
esprima  # For JavaScript/TypeScript (or acorn) 